refactor: split AttenuationOutput in child outputs (wip) [skip ci]
