+++ noisemodelling-jdbc/src/main/java/org/noise_planet/noisemodelling/jdbc/output/NoiseMapWriter.java
@@ -175,7 +175,7 @@ public class NoiseMapWriter implements Callable<Boolean> {
             if(exportPeriod) {
                 ps.setString(parameterIndex++, row.getTimePeriod());
             }
-            ps.setBoolean(parameterIndex++, row.getCutProfile().isFavourable());
+            ps.setBoolean(parameterIndex++, row.isFavourable());
             ps.addBatch();
             batchSize++;
             if (batchSize >= BATCH_MAX_SIZE) {
