+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/profilebuilder/CutProfile.java
@@ -52,7 +52,6 @@ public class CutProfile {
     public PROFILE_TYPE profileType = PROFILE_TYPE.DIRECT;
 
     public Orientation raySourceReceiverDirectivity = new Orientation(); // direction of the source->receiver path relative to the source heading
-    private boolean favourable;
 
     /**
      * Empty constructor for deserialization
@@ -92,14 +91,6 @@ public class CutProfile {
      */
     public boolean isCurvedPath() {
         return curvedPath;
-    }
-
-    public boolean isFavourable() {
-        return favourable;
-    }
-
-    public void setFavourable(boolean favourable) {
-        this.favourable = favourable;
     }
 
     /**
