+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/AttenuationOutput.java
@@ -12,8 +12,6 @@ package org.noise_planet.noisemodelling.propagation;
 
 import org.locationtech.jts.geom.LineString;
 import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
-import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
-import org.noise_planet.noisemodelling.propagation.cnossos.CnossosPath;
 
 /**
  * Output of the attenuation computation (data class).
@@ -24,6 +22,7 @@ public class AttenuationOutput {
     public CutProfile cutProfile;
     public String timePeriod=""; // time period if relevant (day, evening, night or other parameters, use LDenConfig.TIME_PERIOD)
     public LineString lineString; // ray from src to rcv through potential reflexion and diffraction points
+    private boolean favourable;
 
     /**
      * Final attenuation (dB)
@@ -46,6 +45,7 @@ public class AttenuationOutput {
         this.aGlobal = other.aGlobal;
         this.timePeriod = other.timePeriod;
         this.lineString = other.lineString;
+        this.favourable = other.favourable;
     }
 
     /**
@@ -87,5 +87,13 @@ public class AttenuationOutput {
 
     public void setLineString(LineString lineString) {
         this.lineString = lineString;
+    }
+
+    public boolean isFavourable() {
+        return favourable;
+    }
+
+    public void setFavourable(boolean favourable) {
+        this.favourable = favourable;
     }
 }
