+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/cnossos/AttenuationCnossos.java
@@ -12,7 +12,6 @@ package org.noise_planet.noisemodelling.propagation.cnossos;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.math.Vector3D;
 import org.noise_planet.noisemodelling.pathfinder.utils.geometry.Orientation;
-import org.noise_planet.noisemodelling.propagation.cnossos.CnossosAttenuationOutput;
 import org.noise_planet.noisemodelling.propagation.AttenuationParameters;
 import org.noise_planet.noisemodelling.propagation.SceneWithAttenuation;
 import org.slf4j.Logger;
@@ -767,7 +766,7 @@ public class AttenuationCnossos {
             // Favourable conditions
             if (data.getWindRose()[roseIndex] != 0) {
                 cnossosPath.setFavourable(true);
-                attenuationOutput.getCutProfile().setFavourable(true);
+                attenuationOutput.setFavourable(true);
                 aBoundary = AttenuationCnossos.aBoundary(cnossosPath, attenuationOutput, data);
                 aRetroDiff = AttenuationCnossos.deltaRetrodif(cnossosPath, data);
                 for (int idfreq = 0; idfreq < data.getFrequencies().size(); idfreq++) {
